+++ firmware/hdl/packages/lhc_data_pkg.vhd
@@ -2,6 +2,7 @@
 -- Package for definitions of lhc data.
 
 -- Version history:
+-- HB 2023-06-23: inserted calo anomaly algorithm (CICADA) data on link 11 (removed zdc10g on link 11).
 -- HB 2023-02-09: updated comments.
 -- HB 2022-10-08: zdc10g on link 11.
 -- HB 2022-09-08: cleaned up.
@@ -24,11 +25,12 @@
 -- tau                 6 (5..0)        1        8
 -- tau                 6 (11..6)       1        9
 -- esums               6               1       10
--- zdc                 6               1       11
+-- cicada              6               1       11
 -- ext-cond(63..0)     2               1       12
 -- ext-cond(127..64)   2               1       13
 -- ext-cond(191..128)  2               1       14
 -- ext-cond(255..192)  2               1       15
+-- CICADA              6               1       16
 --                     (64 bits)
 -- muon                2 (1..0)        1        0
 -- muon                2 (3..2)        1        1
@@ -46,7 +48,7 @@
 -- tau         5      4     3    2    1      0
 -- tau         11     10    9    8    7      6
 -- esums       HTmHF  ETmHF HTm* ETm* HT,TC* ET,ETTEM*
--- zdc (10G)   5      4     3    2    1      0     
+-- CICADA      5      4     3    2    1      0
 -- ext-cond    x      x     x    x    1      0
 -- ext-cond    x      x     x    x    3      2
 -- ext-cond    x      x     x    x    5      4
@@ -76,6 +78,7 @@ package lhc_data_pkg is
 -- HB 2106-05-31: proposal for memory structure with all frames of calo links for extended test-vector-file structure (see lhc_data_pkg_all_frames.vhd)
     constant OFFSET_ZDC10G_LANES : natural := 11;
     constant OFFSET_EXT_COND_LANES : natural := 12;
+    constant OFFSET_CICADA_LANES : natural := 13;
 
 -- HB 2022-10-10: lane number of ZDC 5G optical input
     constant ZDC5G_LANE_NR : natural := 71;
@@ -97,12 +100,8 @@ package lhc_data_pkg is
     constant HTM_DATA_WIDTH : integer := SW_DATA_WIDTH;
     constant ETMHF_DATA_WIDTH : integer := SW_DATA_WIDTH;
     constant HTMHF_DATA_WIDTH : integer := SW_DATA_WIDTH;
-    constant ZDC10G_0_WIDTH : integer := SW_DATA_WIDTH;
-    constant ZDC10G_1_WIDTH : integer := SW_DATA_WIDTH;
-    constant ZDC10G_2_WIDTH : integer := SW_DATA_WIDTH;
-    constant ZDC10G_3_WIDTH : integer := SW_DATA_WIDTH;
-    constant ZDC10G_4_WIDTH : integer := SW_DATA_WIDTH;
-    constant ZDC10G_5_WIDTH : integer := SW_DATA_WIDTH;
+    constant CICADA_ARRAY_LENGTH : integer := 6;
+    constant CICADA_DATA_WIDTH : integer := SW_DATA_WIDTH;
     constant EXTERNAL_CONDITIONS_DATA_WIDTH : integer := SW_DATA_WIDTH*8;
 
     constant LHC_DATA_WIDTH : integer :=
@@ -117,9 +116,7 @@ package lhc_data_pkg is
             HTM_DATA_WIDTH +
             ETMHF_DATA_WIDTH +
             HTMHF_DATA_WIDTH +
-            ZDC10G_0_WIDTH + ZDC10G_1_WIDTH +
-            ZDC10G_2_WIDTH + ZDC10G_3_WIDTH +
-            ZDC10G_4_WIDTH + ZDC10G_5_WIDTH +
+            (CICADA_ARRAY_LENGTH*CICADA_DATA_WIDTH) +
             EXTERNAL_CONDITIONS_DATA_WIDTH
         );
 
@@ -127,6 +124,7 @@ package lhc_data_pkg is
     type eg_array_t is array(0 to EG_ARRAY_LENGTH-1) of std_logic_vector(EG_DATA_WIDTH-1 downto 0);
     type tau_array_t is array(0 to TAU_ARRAY_LENGTH-1) of std_logic_vector(TAU_DATA_WIDTH-1 downto 0);
     type jet_array_t is array(0 to JET_ARRAY_LENGTH-1) of std_logic_vector(JET_DATA_WIDTH-1 downto 0);
+    type cicada_array_t is array(0 to CICADA_ARRAY_LENGTH-1) of std_logic_vector(CICADA_DATA_WIDTH-1 downto 0);
 
     type lhc_data_t is record
         muon : muon_array_t;
@@ -139,12 +137,7 @@ package lhc_data_pkg is
         htm : std_logic_vector(HTM_DATA_WIDTH-1 downto 0);
         etmhf : std_logic_vector(ETMHF_DATA_WIDTH-1 downto 0);
         htmhf : std_logic_vector(HTMHF_DATA_WIDTH-1 downto 0);
-        zdc10g_0 : std_logic_vector(ZDC10G_0_WIDTH-1 downto 0);
-        zdc10g_1 : std_logic_vector(ZDC10G_1_WIDTH-1 downto 0);
-        zdc10g_2 : std_logic_vector(ZDC10G_2_WIDTH-1 downto 0);
-        zdc10g_3 : std_logic_vector(ZDC10G_3_WIDTH-1 downto 0);
-        zdc10g_4 : std_logic_vector(ZDC10G_4_WIDTH-1 downto 0);
-        zdc10g_5 : std_logic_vector(ZDC10G_5_WIDTH-1 downto 0);
+        cicada : cicada_array_t;
         external_conditions : std_logic_vector(EXTERNAL_CONDITIONS_DATA_WIDTH-1 downto 0);
     end record;
 
@@ -160,16 +153,11 @@ package lhc_data_pkg is
             htm => (others=>'0'),
             etmhf => (others=>'0'),
             htmhf => (others=>'0'),
-            zdc10g_0 => (others=>'0'),
-            zdc10g_1 => (others=>'0'),
-            zdc10g_2 => (others=>'0'),
-            zdc10g_3 => (others=>'0'),
-            zdc10g_4 => (others=>'0'),
-            zdc10g_5 => (others=>'0'),
+            cicada => (others=>(others=>'0')),
             external_conditions => (others=>'0')
         );
 
-    constant LHC_DATA_OBJECT_COUNT : integer :=17;
+    constant LHC_DATA_OBJECT_COUNT : integer :=12;
     constant INDEX_MUON : integer := 0;
     constant INDEX_EG : integer := 1;
     constant INDEX_TAU : integer := 2;
@@ -180,13 +168,9 @@ package lhc_data_pkg is
     constant INDEX_HTM : integer := 7;
     constant INDEX_ETMHF : integer := 8;
     constant INDEX_HTMHF : integer := 9;
-    constant INDEX_ZDC10G_0 : integer := 10;
-    constant INDEX_ZDC10G_1 : integer := 11;
-    constant INDEX_ZDC10G_2 : integer := 12;
-    constant INDEX_ZDC10G_3 : integer := 13;
-    constant INDEX_ZDC10G_4 : integer := 14;
-    constant INDEX_ZDC10G_5 : integer := 15;
-    constant INDEX_EXTERNAL_CONDITIONS : integer := 16;
+    constant INDEX_CICADA : integer := 10;
+    constant INDEX_EXTERNAL_CONDITIONS : integer := 11;
+    
     type lhc_data_slv_property_t is array (0 to LHC_DATA_OBJECT_COUNT-1) of natural;
 
     constant LHC_DATA_SLV_OBJECT_WIDTH : lhc_data_slv_property_t :=
@@ -201,9 +185,7 @@ package lhc_data_pkg is
             HTM_DATA_WIDTH,
             ETMHF_DATA_WIDTH,
             HTMHF_DATA_WIDTH,
-            ZDC10G_0_WIDTH, ZDC10G_1_WIDTH,
-            ZDC10G_2_WIDTH, ZDC10G_3_WIDTH,
-            ZDC10G_4_WIDTH, ZDC10G_5_WIDTH,
+            CICADA_ARRAY_LENGTH * CICADA_DATA_WIDTH,         
             EXTERNAL_CONDITIONS_DATA_WIDTH
         );
 
@@ -259,21 +241,14 @@ package body lhc_data_pkg is
         ret_value(index + HTMHF_DATA_WIDTH-1 downto index) := data_in.htmhf;
         index := index + HTMHF_DATA_WIDTH;
 
-        ret_value(index + ZDC10G_0_WIDTH-1 downto index) := data_in.zdc10g_0;
-        index := index + ZDC10G_0_WIDTH;
-        ret_value(index + ZDC10G_1_WIDTH-1 downto index) := data_in.zdc10g_1;
-        index := index + ZDC10G_1_WIDTH;
-        ret_value(index + ZDC10G_2_WIDTH-1 downto index) := data_in.zdc10g_2;
-        index := index + ZDC10G_2_WIDTH;
-        ret_value(index + ZDC10G_3_WIDTH-1 downto index) := data_in.zdc10g_3;
-        index := index + ZDC10G_3_WIDTH;
-        ret_value(index + ZDC10G_4_WIDTH-1 downto index) := data_in.zdc10g_4;
-        index := index + ZDC10G_4_WIDTH;
-        ret_value(index + ZDC10G_5_WIDTH-1 downto index) := data_in.zdc10g_5;
-        index := index + ZDC10G_5_WIDTH;
+        for i in 0 to CICADA_ARRAY_LENGTH-1 loop
+            ret_value(index+(i+1)*CICADA_DATA_WIDTH-1 downto index+(i*CICADA_DATA_WIDTH)) := data_in.cicada(i);
+        end loop;
+        index := index + (CICADA_ARRAY_LENGTH * CICADA_DATA_WIDTH);
 
         ret_value(index + EXTERNAL_CONDITIONS_DATA_WIDTH-1 downto index) := data_in.external_conditions;
         index := index + EXTERNAL_CONDITIONS_DATA_WIDTH;
+
         return ret_value;
     end function;
 
@@ -316,21 +291,14 @@ package body lhc_data_pkg is
         ret_value.htmhf := data_in(index + HTMHF_DATA_WIDTH-1 downto index);
         index := index + HTMHF_DATA_WIDTH;
 
-        ret_value.zdc10g_0 := data_in(index + ZDC10G_0_WIDTH-1 downto index);
-        index := index + ZDC10G_0_WIDTH;
-        ret_value.zdc10g_1 := data_in(index + ZDC10G_1_WIDTH-1 downto index);
-        index := index + ZDC10G_1_WIDTH;
-        ret_value.zdc10g_2 := data_in(index + ZDC10G_2_WIDTH-1 downto index);
-        index := index + ZDC10G_2_WIDTH;
-        ret_value.zdc10g_3 := data_in(index + ZDC10G_3_WIDTH-1 downto index);
-        index := index + ZDC10G_3_WIDTH;
-        ret_value.zdc10g_4 := data_in(index + ZDC10G_4_WIDTH-1 downto index);
-        index := index + ZDC10G_4_WIDTH;
-        ret_value.zdc10g_5 := data_in(index + ZDC10G_5_WIDTH-1 downto index);
-        index := index + ZDC10G_5_WIDTH;
+        for i in 0 to CICADA_ARRAY_LENGTH-1 loop
+            ret_value.jet(i) := data_in( index+(i+1)*CICADA_DATA_WIDTH-1 downto index+(i* CICADA_DATA_WIDTH));
+        end loop;
+        index := index + (CICADA_ARRAY_LENGTH * CICADA_DATA_WIDTH);
 
         ret_value.external_conditions := data_in(index + EXTERNAL_CONDITIONS_DATA_WIDTH-1 downto index);
         index := index + EXTERNAL_CONDITIONS_DATA_WIDTH;
+
         return ret_value;
     end function;