+++ firmware/hdl/payload/gtl/correlation_conditions.vhd
@@ -3,6 +3,7 @@
 -- Correlation Condition module (for all possible correlation conditions)
 
 -- Version history:
+-- HB 2023-02-03: updated for CICADA.
 -- HB 2022-11-16: bug fix in "mass_3_obj_i/matrix_corr_cond_i": added missing generic parameter.
 -- HB 2022-09-05: cleaned up.
 -- HB 2021-12-09: updated for DISP of jets.
@@ -51,12 +52,18 @@ entity correlation_conditions is
         iso_lut_obj1: std_logic_vector(2**MAX_ISO_BITS-1 downto 0) := (others => '1');
         disp_cut_obj1: boolean := false;
         disp_requ_obj1: boolean := false;
+        bjet_flag_requ_obj1 : boolean := false;
         requested_charge_obj1: string(1 to 3) := "ign";
         qual_lut_obj1: std_logic_vector(2**(MUON_QUAL_HIGH-MUON_QUAL_LOW+1)-1 downto 0) := (others => '1');
         upt_cut_obj1: boolean := false;
         upt_upper_limit_obj1: std_logic_vector(MAX_TEMPLATES_BITS-1 downto 0) := (others => '0');
         upt_lower_limit_obj1: std_logic_vector(MAX_TEMPLATES_BITS-1 downto 0) := (others => '0');
         ip_lut_obj1: std_logic_vector(2**(MUON_IP_HIGH-MUON_IP_LOW+1)-1 downto 0) := (others => '1');
+        hi_bits_requ : boolean := false;
+        hi_bits_val : std_logic_vector(HI_BITS-1 downto 0) := (others => '0');
+        ad_requ : boolean := false;
+        ad_dec_thr : std_logic_vector(AD_DEC_BITS-1 downto 0) := (others => '0');
+        ad_int_thr : std_logic_vector(AD_INT_BITS-1 downto 0) := (others => '0');
 
         slice_low_obj2: natural := 0;
         slice_high_obj2: natural := NR_EG_OBJECTS-1;
@@ -92,6 +99,7 @@ entity correlation_conditions is
         iso_lut_obj2: std_logic_vector(2**MAX_ISO_BITS-1 downto 0) := (others => '1');
         disp_cut_obj2: boolean := false;
         disp_requ_obj2: boolean := false;
+        bjet_flag_requ_obj2 : boolean := false;
         requested_charge_obj2: string(1 to 3) := "ign";
         qual_lut_obj2: std_logic_vector(2**(MUON_QUAL_HIGH-MUON_QUAL_LOW+1)-1 downto 0) := (others => '1');
         upt_cut_obj2: boolean := false;
@@ -133,6 +141,7 @@ entity correlation_conditions is
         iso_lut_obj3: std_logic_vector(2**MAX_ISO_BITS-1 downto 0) := (others => '1');
         disp_cut_obj3: boolean := false;
         disp_requ_obj3: boolean := false;
+        bjet_flag_requ_obj3 : boolean := false;
         requested_charge_obj3: string(1 to 3) := "ign";
         qual_lut_obj3: std_logic_vector(2**(MUON_QUAL_HIGH-MUON_QUAL_LOW+1)-1 downto 0) := (others => '1');
         upt_cut_obj3: boolean := false;
@@ -213,6 +222,9 @@ entity correlation_conditions is
         ls_charcorr_triple: in std_logic_3dim_array(0 to NR_MU_OBJECTS-1, 0 to NR_MU_OBJECTS-1, 0 to NR_MU_OBJECTS-1) := (others => (others => (others => '0')));
         os_charcorr_triple: in std_logic_3dim_array(0 to NR_MU_OBJECTS-1, 0 to NR_MU_OBJECTS-1, 0 to NR_MU_OBJECTS-1) := (others => (others => (others => '0')));
         esums: in std_logic_vector(MAX_ESUMS_BITS-1 downto 0) := (others => '0');
+        hi_bits_i : in std_logic_vector(HI_BITS-1 downto 0) := (others => '0');
+        ad_dec_i : in std_logic_vector(AD_DEC_BITS-1 downto 0) := (others => '0');
+        ad_int_i : in std_logic_vector(AD_INT_BITS-1 downto 0) := (others => '0');        
         deta_orm: in deta_dphi_vector_array(0 to nr_obj1-1, 0 to nr_obj3-1) := (others => (others => (others => '0')));
         dphi_orm: in deta_dphi_vector_array(0 to nr_obj1-1, 0 to nr_obj3-1) := (others => (others => (others => '0')));
         dr_orm: in dr_dim2_array(0 to nr_obj1-1, 0 to nr_obj3-1) := (others => (others => (others => '0')));
@@ -246,7 +258,20 @@ architecture rtl of correlation_conditions is
     signal mass_3_obj_comp_pipe : std_logic_3dim_array(slice_low_obj1 to slice_high_obj1, slice_low_obj2 to slice_high_obj2, slice_low_obj3 to slice_high_obj3) := (others => (others => (others => '0')));
     signal condition_and_or : std_logic;
 
+    signal ad_comp_pipe, hi_comp_pipe : std_logic := '1';
+
 begin
+
+    -- CICADA Anomaly Detection and Heavy Ion Bits comparators
+    cicada_if: if (type_obj1 = BJET_TYPE) or (type_obj2 = BJET_TYPE) or (type_obj3 = BJET_TYPE) generate
+        cicada_i: entity work.cicada_ad_hi_comp
+            generic map(hi_bits_requ, hi_bits_val, ad_requ, ad_dec_thr, ad_int_thr)
+            port map(
+                lhc_clk,
+                hi_bits_i, ad_dec_i, ad_int_i,
+                hi_comp_pipe, ad_comp_pipe
+            );
+    end generate cicada_if;
 
     calo_obj1_sel: if type_obj1 /= MU_TYPE generate
         obj1_l: for i in slice_low_obj1 to slice_high_obj1 generate
@@ -273,7 +298,8 @@ begin
                     phi_w2_lower_limit_obj1,
                     iso_lut_obj1,
                     disp_cut_obj1,
-                    disp_requ_obj1
+                    disp_requ_obj1,
+                    bjet_flag_requ_obj1
                 )
                 port map(lhc_clk, calo_obj1(i), obj1_vs_templ_pipe(i,1));
         end generate obj1_l;
@@ -350,7 +376,8 @@ begin
                         phi_w2_lower_limit_obj2,
                         iso_lut_obj2,
                         disp_cut_obj2,
-                        disp_requ_obj2
+                        disp_requ_obj2,
+                        bjet_flag_requ_obj2
                     )
                     port map(lhc_clk, calo_obj2(i), obj2_vs_templ_pipe(i,1));
             end generate obj2_l;
@@ -489,6 +516,8 @@ begin
                     mass_div_dr_comp => mass_div_dr_comp_pipe,
                     tbpt_comp => tbpt_comp_pipe,
                     charge_comp_double => charge_comp_double_pipe,
+                    hi_comp => hi_comp_pipe,
+                    ad_comp => ad_comp_pipe,
                     condition_and_or => condition_and_or
                 );
 
@@ -522,7 +551,8 @@ begin
                             phi_w2_lower_limit_obj3,
                             iso_lut_obj3,
                             disp_cut_obj3,
-                            disp_requ_obj3
+                            disp_requ_obj3,
+                            bjet_flag_requ_obj3
                         )
                         port map(lhc_clk, calo_obj3(i), obj3_vs_templ_pipe(i,1));
                 end generate obj3_l;
@@ -664,6 +694,8 @@ begin
                         dphi_orm_comp_23 => dphi_orm_comp_23_pipe,
                         dr_orm_comp_13 => dr_orm_comp_13_pipe,
                         dr_orm_comp_23 => dr_orm_comp_23_pipe,
+                        hi_comp => hi_comp_pipe,
+                        ad_comp => ad_comp_pipe,
                         condition_and_or => condition_and_or
                     );
 
@@ -723,6 +755,8 @@ begin
                         deta_orm_comp_12 => deta_orm_comp_12_pipe,
                         dphi_orm_comp_12 => dphi_orm_comp_12_pipe,
                         dr_orm_comp_12 => dr_orm_comp_12_pipe,
+                        hi_comp => hi_comp_pipe,
+                        ad_comp => ad_comp_pipe,
                         condition_and_or => condition_and_or
                     );
 
@@ -786,6 +820,8 @@ begin
                     obj3_vs_templ => obj3_vs_templ_pipe,
                     mass_3_obj_comp => mass_3_obj_comp_pipe,
                     charge_comp_triple => charge_comp_triple_pipe,
+                    hi_comp => hi_comp_pipe,
+                    ad_comp => ad_comp_pipe,
                     condition_and_or => condition_and_or
                 );