+++ firmware/hdl/payload/gtl_module_tpl.vhd
@@ -2,6 +2,7 @@
 -- Global Trigger Logic module.
 
 -- Version history:
+-- HB 2023-06-23: v1.19.0: Implemented logic for CICADA (bjets) on link 11.
 -- HB 2023-04-15: v1.18.1: Bug fix muon index bits.
 -- HB 2023-03-14: v1.18.0: Implemented "MUS2" (new hadronic shower bit), Anomaly Detection Trigger, ZDC conditions and cuts for muon index bits.
 -- HB 2022-11-29: v1.17.6: Bug fix in algo_pre_scaler_fractional_float.vhd.